update progress bar
