+++ .byebug_history
@@ -1,4 +1,14 @@
 exit
+section.count
+exit
+driver.get('https://goodreads.com')
+driver
+exit
+GReadsScraper.new.start_browser
+j
+all_books
+driver
+exit
 rating.avg_rating
 rating
 exit
@@ -244,13 +254,3 @@ book[:time]
 book[:title]
 continue
 book[:title]
-book[:description]
-book[:time]
-book[:author]
-book[author]
-exit
-job[:title]
-exit
-job[:url]
-job[:company]
-job[:location]
