+++ lib/gr_scraper.rb
@@ -5,13 +5,6 @@ class GReadsScraper
     @driver = driver
   end
 
-  def start_browser
-    # options = Selenium::WebDriver::Chrome::Options.new(args: ['headless'])
-    @driver = Selenium::WebDriver.for :chrome
-    # @driver = Watir::driver.new(:chrome)#, switches: %w[--no-sandbox --headless]
-    @driver.get('https://goodreads.com')
-  end
-
   def html
     Nokogiri::HTML.parse(scrap_goodreads.page_source)
   end
@@ -26,15 +19,5 @@ class GReadsScraper
   def scrap_goodreads
     @driver.find_element(id: 'sitesearch_field').send_keys @all_books[@title_index][1], :return
     @driver
-  end
-end
-
-class ScrapNYT
-  def scrap_nytimes
-    Nokogiri::HTML(open('https://www.nytimes.com/books/best-sellers/'))
-  end
-
-  def scrap_sections
-    scrap_nytimes.css('section.e8j42380')
   end
 end
